+++ TombEngine/Game/effects/effects.h
@@ -283,6 +283,7 @@ short DoBloodSplat(int x, int y, int z, short speed, short yRot, short roomNumbe
 void DoLotsOfBlood(int x, int y, int z, int speed, short direction, short roomNumber, int count);
 void ControlWaterfallMist(short itemNumber);
 void TriggerWaterfallMist(const ItemInfo& item);
+void TriggerWaterfallMist(Vector3 pos, int size, int width, float angle, Vector4 color);
 void KillAllCurrentItems(short itemNumber);
 void TriggerRocketFlame(int x, int y, int z, int xv, int yv, int zv, int itemNumber);
 void TriggerRocketSmoke(int x, int y, int z);
@@ -291,7 +292,7 @@ void TriggerMetalSparks(int x, int y, int z, int xv, int yv, int zv, const Vecto
 void SpawnCorpseEffect(const Vector3& pos);
 void TriggerAttackFlame(const Vector3i& pos, const Vector3& color, int scale);
 void TriggerRocketFire(int x, int y, int z);
-void TriggerExplosionBubbles(int x, int y, int z, short roomNumber);
+void TriggerExplosionBubbles(int x, int y, int z, short roomNumber, const Vector3& mainColor = Vector3::Zero, const Vector3& secondColor = Vector3::Zero);
 void Ricochet(Pose& pos);
 void ProcessEffects(ItemInfo* item);
 void UpdateWibble();