+++ TombEngine/Game/effects/tomb4fx.h
@@ -300,7 +300,7 @@ extern SHOCKWAVE_STRUCT ShockWaves[MAX_SHOCKWAVE];
 extern std::vector<FIRE_LIST> Fires;
 
 void TriggerBlood(int x, int y, int z, int unk, int num);
-void TriggerExplosionBubble(int x, int y, int z, short roomNumber);
+void TriggerExplosionBubble(int x, int y, int z, short roomNumber, const Vector3& mainColor = Vector3::Zero, const Vector3& secondColor = Vector3::Zero);
 int GetFreeFireSpark();
 void TriggerGlobalStaticFlame();
 void TriggerGlobalFireSmoke();
@@ -332,5 +332,6 @@ void TriggerShockwave(Pose* pos, short innerRad, short outerRad, int speed, unsi
 void TriggerShockwaveHitEffect(int x, int y, int z, unsigned char r, unsigned char g, unsigned char b, short rot, int vel);
 void UpdateShockwaves();
 void TriggerSmallSplash(int x, int y, int z, int number);
-void TriggerUnderwaterExplosion(ItemInfo* item, int flag);
+void TriggerUnderwaterExplosion(ItemInfo* item, bool flag);
+void TriggerUnderwaterExplosion(Vector3 position, bool flag, const Vector3& mainColor = Vector3::Zero, const Vector3& secondColor = Vector3::Zero);
 void ExplodeVehicle(ItemInfo* laraItem, ItemInfo* vehicle);