+++ TombEngine/Scripting/Internal/TEN/Effects/EffectsFunctions.cpp
@@ -4,6 +4,7 @@
 #include "Game/camera.h"
 #include "Game/collision/collide_room.h"
 #include "Game/control/los.h"
+#include "Game/effects/blood.h"
 #include "Game/effects/Bubble.h"
 #include "Game/effects/DisplaySprite.h"
 #include "Game/effects/effects.h"
@@ -32,10 +33,12 @@
 #include "Specific/trutils.h"
 #include <Scripting/Internal/TEN/Objects/Moveable/MoveableObject.h>
 
+
 /// Functions to generate effects.
 // @tentable Effects 
 // @pragma nostrip
 
+using namespace TEN::Effects::Blood;
 using namespace TEN::Effects::Bubble;
 using namespace TEN::Effects::DisplaySprite;
 using namespace TEN::Effects::Electricity;
@@ -477,11 +480,16 @@ namespace TEN::Scripting::Effects
 
 	/// Emit blood.
 	// @function EmitBlood
-	// @tparam Vec3 pos
-	// @tparam[opt=1] int count Sprite count.
+	// @tparam Vec3 pos World position.
+	// @tparam[opt=1] int count Blood count.
 	static void EmitBlood(const Vec3& pos, TypeOrNil<int> count)
 	{
-		TriggerBlood(pos.x, pos.y, pos.z, -1, ValueOr<int>(count, 1));
+		int roomNumber = FindRoomNumber(pos.ToVector3i());
+		const auto& room = g_Level.Rooms[roomNumber];
+		if (room.flags & ENV_FLAG_WATER)
+			SpawnUnderwaterBloodCloud(pos, roomNumber, (GetRandomControl() & 7) + 8, ValueOr<int>(count, 1));
+		else
+			TriggerBlood(pos.x, pos.y, pos.z, -1, ValueOr<int>(count, 1));
 	}
 
 	/// Emit an air bubble in a water room.
@@ -500,6 +508,27 @@ namespace TEN::Scripting::Effects
 		SpawnBubble(pos.ToVector3(), roomNumber, convertedSize, convertedAmp);
 	}
 
+	/// Emit waterfall mist.
+	// @function EmitWaterfallMist
+	// @tparam Vec3 pos World position where the effect will be spawned.
+	// @tparam[opt=64] float size Effect size.
+	// @tparam[opt=32] float width Width of the effect.
+	// @tparam[opt=0] float rot Rotation of effect in degrees.
+	// @tparam[opt=Color(255&#44; 255&#44; 255&#44; 255)] Color color Color of the effect.
+	static void EmitWaterfallMist(const Vec3& pos, TypeOrNil<int> size, TypeOrNil<int> width, TypeOrNil<float> angle, TypeOrNil<ScriptColor> color)
+	{
+		constexpr auto DEFAULT_SIZE = 64;
+		constexpr auto DEFAULT_WIDTH = 32;
+
+		auto convertedAngle = ANGLE(ValueOr<float>(angle, 0.0f));
+		auto convertedSize = ValueOr<int>(size, DEFAULT_SIZE);
+		auto convertedWidth = ValueOr<int>(width, DEFAULT_WIDTH);
+		auto _color = ValueOr<ScriptColor>(color, ScriptColor(255, 255, 255));
+		auto convertedColor = Vector4(_color.GetR(), _color.GetG(), _color.GetB(), _color.GetA()) / UCHAR_MAX;
+
+		TriggerWaterfallMist(pos, convertedSize, convertedWidth, convertedAngle, convertedColor);
+	}
+
 	/// Emit fire for one frame. Will not hurt player. Call this each frame if you want a continuous fire.
 	// @function EmitFire
 	// @tparam Vec3 pos Position.
@@ -511,12 +540,39 @@ namespace TEN::Scripting::Effects
 
 	/// Make an explosion. Does not hurt Lara
 	// @function MakeExplosion 
-	// @tparam Vec3 pos
-	// @tparam[opt=512] float size This will not be the size of the sprites, but rather the distance between the origin and any additional sprites.
-	// @tparam[opt=false] bool shockwave If true, create a very faint white shockwave which will not hurt Lara.
-	static void MakeExplosion(Vec3 pos, TypeOrNil<float> size, TypeOrNil<bool> shockwave)
+	// @tparam Vec3 pos World position.
+	// @tparam[opt=512] float size Size of the shockwave if enabled.
+	// @tparam[opt=false] bool shockwave If true, create a very faint shockwave which will not hurt Lara. For underwater rooms it creates a splash if pos is near the surface. Shockwave uses Main color if provided.
+	// @tparam[opt] Color maincol Main Color of the explosion and the shockwave. If not provided default explosion color will be used.
+	// @tparam[opt] Color secondCol Second Color of the explosion. If not provided default explosion color will be used.
+	static void MakeExplosion(Vec3 pos, TypeOrNil<float> size, TypeOrNil<bool> shockwave, TypeOrNil<ScriptColor> mainCol, TypeOrNil<ScriptColor> secondCol)
 	{
-		TriggerExplosion(Vector3(pos.x, pos.y, pos.z), ValueOr<float>(size, 512.0f), true, false, ValueOr<bool>(shockwave, false), FindRoomNumber(Vector3i(pos.x, pos.y, pos.z)));
+		auto convertedShockwave = ValueOr<bool>(shockwave, false);
+		auto convertedSize = ValueOr<float>(size, 512.0f);
+
+		auto convertedMainColor = ValueOr<ScriptColor>(mainCol, ScriptColor(0, 0, 0));
+		auto convertedSecondColor = ValueOr<ScriptColor>(secondCol, convertedMainColor);
+
+		int roomNumber = FindRoomNumber(pos.ToVector3i());
+		const auto& room = g_Level.Rooms[roomNumber];
+
+		if (room.flags & ENV_FLAG_WATER)
+			TriggerUnderwaterExplosion(pos.ToVector3(), ValueOr<bool>(shockwave, false), Vector3(convertedMainColor), Vector3(convertedSecondColor));
+		else
+		{
+			TriggerExplosionSparks(pos.x, pos.y, pos.z, 3, -2, 0, FindRoomNumber(Vector3i(pos.x, pos.y, pos.z)), Vector3(convertedMainColor), Vector3(convertedSecondColor));
+
+			if (convertedShockwave)
+			{
+				auto shockPos = Pose(Vector3i(pos));
+
+				if (Vector3(convertedMainColor) == Vector3::Zero)
+				TriggerShockwave(&shockPos, 0, convertedSize, 64, 128, 96, 0, 30, EulerAngles(rand() & 0xFFFF, 0.0f, 0.0f), 0, true, false, false, (int)ShockwaveStyle::Normal);
+				else
+				TriggerShockwave(&shockPos, 0, convertedSize, 64, convertedMainColor.GetR(), convertedMainColor.GetG(), convertedMainColor.GetB(), 30, EulerAngles(rand() & 0xFFFF, 0.0f, 0.0f), 0, true, false, false, (int)ShockwaveStyle::Normal);
+			}
+
+		}
 	}
 
 	/// Make an earthquake
@@ -582,6 +638,84 @@ namespace TEN::Scripting::Effects
 			convertedEdgeFeatherID, /*convertedLengthFeatherID, */convertedBlendID);
 	}
 
+	/// Emit a particle flowing effect.
+	// @function EmitFlow
+	// @tparam Vec3 pos World position.
+	// @tparam Vec3 dir Directional velocity of the particles in world units per second.
+	// @tparam[opt=512] float radius Radius of emitter. The particles will be emitted inside the circle of provided radius measured from centre of world position.
+	// @tparam[opt=1] float life Lifespan in seconds.
+	// @tparam[opt=0] float friction Friction affecting velocity over time in world units per second.
+	// @tparam[opt=25] float maxSize Max size of the particle.
+	// @tparam[opt=Color(128&#44; 128&#44; 128)] Color startColor Color at start of life.
+	// @tparam[opt=Color(0&#44; 0&#44; 0)] Color endColor Color at end of life.
+	// @tparam[opt=Objects.ObjID.DEFAULT_SPRITES] Objects.ObjID.SpriteConstants spriteSeqID Sprite sequence slot ID.
+	// @tparam[opt=14 (UNDERWATER_DUST)] int spriteID Sprite ID in the sprite sequence slot.
+	static void EmitFlow(const Vec3& pos, const Vec3& dir, TypeOrNil<float> radius, TypeOrNil<float> life, TypeOrNil<float> friction, TypeOrNil<float> maxSize, TypeOrNil<ScriptColor> startColor, TypeOrNil<ScriptColor> endColor, TypeOrNil<GAME_OBJECT_ID> spriteSeqID, TypeOrNil<int> spriteID)
+	{
+		constexpr auto DEFAULT_LIFE = 1.0f;
+		constexpr auto SECS_PER_FRAME = 1.0f / (float)FPS;
+		constexpr auto DUST_SIZE_MAX = 25.0f;
+
+		auto convertedSpriteSeqID = ValueOr<GAME_OBJECT_ID>(spriteSeqID, ID_DEFAULT_SPRITES);
+		if (!CheckIfSlotExists(convertedSpriteSeqID, "EmitParticle() script function."))
+			return;
+
+		auto convertedPos = pos.ToVector3();
+		auto convertedDir = dir.ToVector3() / (float)FPS;
+		auto convertedRad = ValueOr<float>(radius, BLOCK(0.5f));
+		auto convertedLife = std::max(0.1f, ValueOr<float>(life, DEFAULT_LIFE));
+		auto convertedFriction = ValueOr<float>(friction, 0) / (float)FPS;
+		auto convertedMaxSize = std::max(0.1f, ValueOr<float>(maxSize, DUST_SIZE_MAX));
+		auto convertedStartColor = ValueOr<ScriptColor>(startColor, ScriptColor(128, 128, 128, 255));
+		auto convertedEndColor = ValueOr<ScriptColor>(endColor, ScriptColor(0, 0, 0, 255));
+		auto convertedSpriteID = ValueOr<int>(spriteID, SPRITE_TYPES::SPR_UNDERWATERDUST);
+
+		auto& part = *GetFreeParticle();
+
+		part.on = true;
+		part.SpriteSeqID = convertedSpriteSeqID;
+		part.SpriteID = convertedSpriteID;
+		part.blendMode = BlendMode::Additive;
+
+		// Set particle colors
+		part.sR = convertedStartColor.GetR();
+		part.sG = convertedStartColor.GetG();
+		part.sB = convertedStartColor.GetB();
+
+		part.dR = convertedEndColor.GetR();
+		part.dG = convertedEndColor.GetG();
+		part.dB = convertedEndColor.GetB();
+
+		part.life =
+			part.sLife = (int)round(convertedLife / SECS_PER_FRAME);
+		part.colFadeSpeed = part.life / 2;
+		part.fadeToBlack = part.life / 3;
+
+		// Randomize position within the given radius
+		float angle = TO_DEGREES(Random::GenerateAngle());
+		float randRadius = sqrt(Random::GenerateFloat()) * convertedRad;
+
+		part.x = convertedPos.x + randRadius * cos(angle);
+		part.y = convertedPos.y + (Random::GenerateFloat() * 2.0f - 1.0f) * convertedRad;
+		part.z = convertedPos.z + randRadius * sin(angle);
+		part.roomNumber = FindRoomNumber(Vector3i(part.x, part.y, part.z));
+
+		part.xVel = convertedDir.x * 32;
+		part.yVel = convertedDir.y * 32;
+		part.zVel = convertedDir.z * 32;
+
+		part.rotAng = ANGLE(0.0f) >> 4;
+		part.rotAdd = ANGLE(0.0f) >> 4;
+
+		// Other properties
+		part.friction = convertedFriction;
+		part.maxYvel = 0;
+		part.gravity = 0;
+		part.flags = SP_SCALE | SP_ROTATE | SP_DEF | SP_EXPDEF;
+		part.scalar = 2;
+		part.sSize = part.size = part.dSize = Random::GenerateFloat(convertedMaxSize / 2, convertedMaxSize);
+	}
+
 	void Register(sol::state* state, sol::table& parent) 
 	{
 		auto tableEffects = sol::table(state->lua_state(), sol::create);
@@ -598,7 +732,8 @@ namespace TEN::Scripting::Effects
 		tableEffects.set_function(ScriptReserved_EmitAirBubble, &EmitAirBubble);
 		tableEffects.set_function(ScriptReserved_EmitStreamer, &EmitStreamer);
 		tableEffects.set_function(ScriptReserved_EmitFire, &EmitFire);
-
+		tableEffects.set_function(ScriptReserved_EmitWaterfallMist, &EmitWaterfallMist);
+		tableEffects.set_function(ScriptReserved_EmitFlow, &EmitFlow);
 		tableEffects.set_function(ScriptReserved_MakeExplosion, &MakeExplosion);
 		tableEffects.set_function(ScriptReserved_MakeEarthquake, &Earthquake);
 		tableEffects.set_function(ScriptReserved_GetWind, &GetWind);